Fixed formatting

+++ src/lib.rs
@@ -1,30 +1,30 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // See Notices.txt for copyright information
-//#![cfg_attr(not(test), no_std)]
+
+// Disabled no-std mode, it requires core_ffi_c feature stabilized
+// #![cfg_attr(not(test), no_std)]
 #![allow(non_snake_case, non_camel_case_types, non_upper_case_globals)]
 
 #[allow(non_camel_case_types)]
 #[allow(unused)]
-mod internal
-{
-	pub type int8_t = i8;
-	pub type int16_t = i16;
-	pub type int32_t = i32;
-	pub type int64_t = i64;
+mod internal {
+    pub type int8_t = i8;
+    pub type int16_t = i16;
+    pub type int32_t = i32;
+    pub type int64_t = i64;
 
-	pub type uint8_t = u8;
-	pub type uint16_t = u16;
-	pub type uint32_t = u32;
-	pub type uint64_t = u64;
+    pub type uint8_t = u8;
+    pub type uint16_t = u16;
+    pub type uint32_t = u32;
+    pub type uint64_t = u64;
 
     include!(concat!(env!("OUT_DIR"), "/c99-types.rs"));
 }
 
 pub use self::internal::{
-    int8_t, int16_t, int32_t, int64_t,
-    uint8_t, uint16_t, uint32_t, uint64_t,
-    int_fast8_t, int_fast16_t, int_fast32_t, int_fast64_t,
-    uint_fast8_t, uint_fast16_t, uint_fast32_t, uint_fast64_t,
+    int16_t, int32_t, int64_t, int8_t, int_fast16_t, int_fast32_t, int_fast64_t, int_fast8_t,
+    uint16_t, uint32_t, uint64_t, uint8_t, uint_fast16_t, uint_fast32_t, uint_fast64_t,
+    uint_fast8_t,
 };
 
 extern "C" {
